+++ setup.py
@@ -18,17 +18,19 @@ setup(
     ),
     install_requires=["geopandas", "networkx", "pysheds", "rtree"],
     extras_require={
-        "basemap": "contextily",
         "dev": ["ruff", "pytest"],
-        "plotting": "matplotlib",
+        "map": ["contextily", "matplotlib"],
+        "util": ["rasterio"],
     },
     classifiers=[
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
         "Topic :: Scientific/Engineering :: GIS",
         "Topic :: Scientific/Engineering :: Hydrology",
     ],
-    python_requires=">=3.8",
+    # TODO: Enforce this?
+    # python_requires=">=3.10",
 )
