+++ stormcatchments/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.4.1"
 
-from . import constants, delineate, network, terrain, topology # noqa
+from . import delineate, network # noqa
 from .network import Network # noqa
 from .delineate import Delineate # noqa
